Add navigation
Add navigation to switch between different pages on the site, including
a placeholder library tab.

Fixes: #95.
Part-of: #90.
