+++ client/package.json
@@ -13,13 +13,14 @@
     "react-bootstrap": "^2.6.0",
     "react-dom": "^18.2.0",
     "react-force-graph": "^1.41.14",
-    "react-scripts": "^2.1.3"
+    "react-scripts": "^2.1.3",
+    "react-router-dom": "^6.4.3"
   },
   "scripts": {
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build",
-    "start": "react-scripts start --openssl-legacy-provider",
-    "build": "react-scripts build --openssl-legacy-provider",
+    "start": "react-scripts --openssl-legacy-provider start",
+    "build": "react-scripts --openssl-legacy-provider build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
