+++ client/src/Library.js
@@ -1,0 +1,34 @@
+import React, { useEffect } from "react";
+import Card from "react-bootstrap/Card";
+import Col from "react-bootstrap/Col";
+import Row from "react-bootstrap/Row";
+import "./App.css";
+import "./Library.css";
+
+function Library() {
+  useEffect(() => {
+    document.body.style.backgroundColor = "#eae0d5";
+  });
+
+  return (
+    <div className="books">
+      <Row xs={1} md={2} lg={4} className="g-4">
+        {Array.from({ length: 4 }).map((_, idx) => (
+          <Col>
+            <Card>
+              <Card.Img variant="top" src="holder.js/100px160" />
+              <Card.Body>
+                <Card.Title>Book title</Card.Title>
+                <Card.Text>
+                  Give a short description of the data present for this text.
+                </Card.Text>
+              </Card.Body>
+            </Card>
+          </Col>
+        ))}
+      </Row>
+    </div>
+  );
+}
+
+export default Library;
